More detailed tests for UnnecessaryUnicodeEscapeInspection
I had been confused when I saw the `do while` loop in the code since it
looked like an obvious endless loop. Therefore I wrote some tests to
make sure this code is covered (which it already was before).

Since extracting the escape sequences from the source code is
independent of reporting the errors, I extracted that code to its own
method, which made it easily testable on its own.